LPD-51224 Extract the test into a separate class (aka move the orange juice out of the cleaning supplies aisle)
